src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
discord.py==2.5.2
distro==1.9.0
fastapi==0.115.14
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.24.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.6.3
openai==1.93.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pypi==2.1
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
starlette==0.46.2
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1